+++ constants/mutation.ts
@@ -2,3 +2,5 @@ export enum Mutations {
   CREATE_PLAYLIST = "Create Playlist",
   ADD_MOVIE_TO_PLAYLIST = "Add Movie To Playlist",
 }
+
+export const PENDING_MUTATIONS_KEY = "pending_mutations";
