+++ hooks/useNetworkDetection.ts
@@ -1,12 +1,11 @@
 import { useState, useEffect, useRef, useCallback } from "react";
 import NetInfo from "@react-native-community/netinfo";
 import { onlineManager, QueryClient } from "@tanstack/react-query";
-import { useProcessPendingMutations } from "./useProcessPendingMutations";
+import { processPendingMutations } from "./useProcessPendingMutations";
 
 export default function useNetworkDetection(queryClient: QueryClient) {
   const [modalVisible, setModalVisible] = useState(<boolean>false);
   const [message, setMessage] = useState<string>("");
-  const { processPendingMutations } = useProcessPendingMutations(queryClient);
 
   // Track initial connection state to avoid showing modal on app launch
   const hasCheckedInitialConnection = useRef<boolean>(false);
@@ -18,7 +17,7 @@ export default function useNetworkDetection(queryClient: QueryClient) {
       if (isConnected) {
         setMessage("You are back online!");
         queryClient.resumePausedMutations(); // Handle pending mutations (using queryClient)
-        // processPendingMutations(); // Handle pending mutations (ourselves)
+        // processPendingMutations(queryClient); // Handle pending mutations (ourselves)
       } else {
         setMessage("You are offline!");
       }
