+++ types/Mutation.ts
@@ -1,0 +1,6 @@
+export type PendingMutation = {
+  id: string;
+  mutationKey: string;
+  data: any;
+  timestamp: number;
+};
